Added food items per user (not just one master list)

+++ resources/views/welcome.blade.php
@@ -39,7 +39,7 @@
                             <th>Move</th>
                         </tr>
                         @foreach ($items as $item)
-                            @if ($item -> inGroceryList == true)
+                            @if ($item -> inGroceryList == true && $item -> userId == Auth::user()->id)
                                  <tr>
                                     <?php 
                                         $function = 'deleteItem(\''.$item->id.'\')';
@@ -72,7 +72,7 @@
                             <th>Action</th>
                         </tr>
                          @foreach ($items as $item)
-                            @if ($item -> inPantry == true)
+                            @if ($item -> inPantry == true && $item -> userId == Auth::user()->id)
                                  <tr>
                                     <?php 
                                         $function = 'deleteItem(\''.$item->id.'\')';
